requirements_app.txt
streamlit==1.36.0
openai==1.35.10
python-pptx==0.6.23
requests==2.32.3
cloudconvert==2.1.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
requests
fastapi
convertapi